Module emulated_integration_tests_common::impls

Re-exports

pub use crate::xcm_helpers::xcm_transact_paid_execution;

pub use crate::xcm_helpers::xcm_transact_unpaid_execution;

pub use crate::PROOF_SIZE_THRESHOLD;

pub use crate::REF_TIME_THRESHOLD;

pub use paste;

pub use pallet_assets;

pub use pallet_message_queue;

pub use pallet_xcm;

pub use cumulus_pallet_parachain_system;

pub use cumulus_pallet_xcmp_queue;

Modules

dmp

To prevent Out of Memory errors on the DownwardMessageQueue, an exponential fee factor (DeliveryFeeFactor) is set. The fee factor increments exponentially after the number of messages in the DownwardMessageQueue passes a threshold. This threshold is set as:

hrmp

Macros

assert_expected_events

assert_ok

Panic if an expression doesn’t evaluate to Ok.

bx

Structs

AccountId32

An opaque 32-byte cryptographic identifier.

Asset

Either an amount of a single fungible asset, or a single well-identified non-fungible asset.

BridgeHubMessageHandler

BridgeMessage

BridgeMessageDispatchError

BridgeMessagesInstance1

Instance1 to be used for instantiable pallets defined with the #[pallet] macro. Instances 2-16 are also available but are hidden from docs.

DoubleEncoded

Wrapper around the encoded and decoded versions of a value. Caches the decoded value once computed.

HrmpChannelId

A type that uniquely identifies an HRMP channel. An HRMP channel is established between two paras. In text, we use the notation (A, B) to specify a channel between A and B. The channels are unidirectional, meaning that (A, B) and (B, A) refer to different channels. The convention is that we use the first item tuple for the sender and the second for the recipient. Only one channel is allowed between two participants in one direction, i.e. there cannot be 2 different channels identified by (A, B). A channel with the same para id in sender and recipient is invalid. That is, however, not enforced.

Location

A relative path between state-bearing consensus systems.

ParaId

Unique identifier of a parachain.

Weight

WeightMeter

Meters consumed weight and a hard limit for the maximal consumable weight.

Enums

AggregateMessageOrigin

Aggregate message origin for the MessageQueue pallet.

Junction

A single item in a path to describe the relative location of a consensus system.

Junctions

Non-parent junctions that can be constructed, up to the length of 8. This specific Junctions implementation uses a Rust enum in order to make pattern matching easier.

NetworkId

A global identifier of a data structure existing within consensus.

OriginKind

Basically just the XCM (more general) version of ParachainDispatchOrigin.

Outcome

Outcome of an XCM execution.

UmpQueueId

Identifies a UMP queue inside the MessageQueue pallet.

VersionedXcm

A single XCM message, together with its version code.

XcmBridgeHubCall

A minimized version of pallet_xcm_bridge_hub::Call that can be used without a runtime.

XcmError

Error codes used in XCM. The first errors codes have explicit indices and are part of the XCM format. Those trailing are merely part of the XCM implementation; there is no expectation that they will retain the same index over time.

Traits

BridgeMessageHandler

Chain

Decode

Trait that allows zero-copy read of value-references from slices in LE format.

DmpMessageHandler

Something that should be called when a downward message is received.

Encode

Trait that allows zero-copy write of value-references to slices in LE format.

Inspect

Trait for providing balance-inspection access to a set of named fungible assets.

Network

Parachain

RelayChain

TestExt

XcmpMessageHandler

Something that should be called for each batch of messages received over XCMP.

Functions

weight_within_threshold

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Balance

Balance of an account.

BridgeLaneId

InteriorLocation

A relative location which is constrained to be an interior location of the context.

XcmVersion

A version of XCM.

Derive Macros

Decode

Derive parity_scale_codec::Decode and for struct and enum.

Encode

Derive parity_scale_codec::Encode and parity_scale_codec::EncodeLike for struct and enum.