Module emulated_integration_tests_common::impls::hrmp

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

GenesisConfig

Preopen the given HRMP channels.

HrmpChannel

A metadata of an HRMP channel.

HrmpOpenChannelRequest

A description of a request to open an HRMP channel.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

TestWeightInfo

A weight info that is only suitable for testing.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

Constants

HRMP_MAX_INBOUND_CHANNELS_BOUND

Maximum bound that can be set for inbound channels.

HRMP_MAX_OUTBOUND_CHANNELS_BOUND

Same as HRMP_MAX_INBOUND_CHANNELS_BOUND, but for outbound channels.

Traits

Config

Configuration trait of this pallet.

WeightInfo

Type Definitions

HrmpAcceptedChannelRequestCount

This mapping tracks how many open channel requests were accepted by a given recipient para. Invariant: HrmpOpenChannelRequests should contain the same number of items (_, X) with confirmed set to true, as the number of HrmpAcceptedChannelRequestCount for X.

HrmpChannelContents

Storage for the messages for each channel. Invariant: cannot be non-empty if the corresponding channel in HrmpChannels is None.

HrmpChannelDigests

Maintains a mapping that can be used to answer the question: What paras sent a message at the given block number for a given receiver. Invariants:

HrmpChannels

HRMP channel data associated with each para. Invariant:

HrmpCloseChannelRequests

A set of pending HRMP close channel requests that are going to be closed during the session change. Used for checking if a given channel is registered for closure.

HrmpCloseChannelRequestsList

Storage type is StorageValue with value type Vec < HrmpChannelId >.

HrmpEgressChannelsIndex

Storage type is StorageMap with key type ParaId and value type Vec < ParaId >.

HrmpIngressChannelsIndex

Ingress/egress indexes allow to find all the senders and receivers given the opposite side. I.e.

HrmpOpenChannelRequestCount

This mapping tracks how many open channel requests are initiated by a given sender para. Invariant: HrmpOpenChannelRequests should contain the same number of items that has (X, _) as the number of HrmpOpenChannelRequestCount for X.

HrmpOpenChannelRequests

The set of pending HRMP open channel requests.

HrmpOpenChannelRequestsList

Storage type is StorageValue with value type Vec < HrmpChannelId >.

HrmpWatermarks

The HRMP watermark associated with each para. Invariant:

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.