Type Alias emulated_integration_tests_common::impls::hrmp::pallet::Module

pub type Module<T> = Pallet<T>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T> Pallet<T>where T: Config,

pub fn hrmp_init_open_channel( origin: <T as Config>::RuntimeOrigin, recipient: Id, proposed_max_capacity: u32, proposed_max_message_size: u32 ) -> Result<(), DispatchError>

Initiate opening a channel from a parachain to a given recipient with given channel parameters.

proposed_max_capacity - specifies how many messages can be in the channel at once.

proposed_max_message_size - specifies the maximum size of the messages.

These numbers are a subject to the relay-chain configuration limits.

The channel can be opened only after the recipient confirms it and only on a session change.

pub fn hrmp_accept_open_channel( origin: <T as Config>::RuntimeOrigin, sender: Id ) -> Result<(), DispatchError>

Accept a pending open channel request from the given sender.

The channel will be opened only on the next session boundary.

pub fn hrmp_close_channel( origin: <T as Config>::RuntimeOrigin, channel_id: HrmpChannelId ) -> Result<(), DispatchError>

Initiate unilateral closing of a channel. The origin must be either the sender or the recipient in the channel being closed.

The closure can only happen on a session change.

pub fn force_clean_hrmp( origin: <T as Config>::RuntimeOrigin, para: Id, num_inbound: u32, num_outbound: u32 ) -> Result<(), DispatchError>

This extrinsic triggers the cleanup of all the HRMP storage items that a para may have. Normally this happens once per session, but this allows you to trigger the cleanup immediately for a specific parachain.

Number of inbound and outbound channels for para must be provided as witness data.

Origin must be the ChannelManager.

pub fn force_process_hrmp_open( origin: <T as Config>::RuntimeOrigin, channels: u32 ) -> Result<(), DispatchError>

Force process HRMP open channel requests.

If there are pending HRMP open channel requests, you can use this function to process all of those requests immediately.

Total number of opening channels must be provided as witness data.

Origin must be the ChannelManager.

pub fn force_process_hrmp_close( origin: <T as Config>::RuntimeOrigin, channels: u32 ) -> Result<(), DispatchError>

Force process HRMP close channel requests.

If there are pending HRMP close channel requests, you can use this function to process all of those requests immediately.

Total number of closing channels must be provided as witness data.

Origin must be the ChannelManager.

pub fn hrmp_cancel_open_request( origin: <T as Config>::RuntimeOrigin, channel_id: HrmpChannelId, open_requests: u32 ) -> Result<(), DispatchError>

This cancels a pending open channel request. It can be canceled by either of the sender or the recipient for that request. The origin must be either of those.

The cancellation happens immediately. It is not possible to cancel the request if it is already accepted.

Total number of open requests (i.e. HrmpOpenChannelRequestsList) must be provided as witness data.

pub fn force_open_hrmp_channel( origin: <T as Config>::RuntimeOrigin, sender: Id, recipient: Id, max_capacity: u32, max_message_size: u32 ) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>

Open a channel from a sender to a recipient ParaId. Although opened by governance, the max_capacity and max_message_size are still subject to the Relay Chain’s configured limits.

Expected use is when one (and only one) of the ParaIds involved in the channel is governed by the system, e.g. a system parachain.

Origin must be the ChannelManager.

pub fn establish_system_channel( origin: <T as Config>::RuntimeOrigin, sender: Id, recipient: Id ) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>

Establish an HRMP channel between two system chains. If the channel does not already exist, the transaction fees will be refunded to the caller. The system does not take deposits for channels between system chains, and automatically sets the message number and size limits to the maximum allowed by the network’s configuration.

Arguments:

sender: A system chain, ParaId.

recipient: A system chain, ParaId.

Any signed origin can call this function, but both inputs MUST be system chains. If the channel does not exist yet, there is no fee.

pub fn poke_channel_deposits( origin: <T as Config>::RuntimeOrigin, sender: Id, recipient: Id ) -> Result<(), DispatchError>

Update the deposits held for an HRMP channel to the latest Configuration. Channels with system chains do not require a deposit.

Arguments:

sender: A chain, ParaId.

recipient: A chain, ParaId.

Any signed origin can call this function.

impl<T> Pallet<T>where T: Config,

Routines and getters related to HRMP.

pub fn init_open_channel( origin: Id, recipient: Id, proposed_max_capacity: u32, proposed_max_message_size: u32 ) -> Result<(), DispatchError>

Initiate opening a channel from a parachain to a given recipient with given channel parameters. If neither chain is part of the system, then a deposit from the Configuration will be required for origin (the sender) upon opening the request and the recipient upon accepting it.

Basically the same as hrmp_init_open_channel but intended for calling directly from other pallets rather than dispatched.

pub fn accept_open_channel(origin: Id, sender: Id) -> Result<(), DispatchError>

Accept a pending open channel request from the given sender.

Basically the same as hrmp_accept_open_channel but intended for calling directly from other pallets rather than dispatched.