Trait emulated_integration_tests_common::impls::hrmp::Config

pub trait Config: Config + Config + Config + Config {
    type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>;
    type RuntimeOrigin: From<Origin> + From<Self::RuntimeOrigin> + Into<Result<Origin, Self::RuntimeOrigin>>;
    type ChannelManager: EnsureOrigin<Self::RuntimeOrigin>;
    type Currency: ReservableCurrency<Self::AccountId>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>

The outer event type.

type RuntimeOrigin: From<Origin> + From<Self::RuntimeOrigin> + Into<Result<Origin, Self::RuntimeOrigin>>

type ChannelManager: EnsureOrigin<Self::RuntimeOrigin>

The origin that can perform “force” actions on channels.

type Currency: ReservableCurrency<Self::AccountId>

An interface for reserving deposits for opening channels.

NOTE that this Currency instance will be charged with the amounts defined in the Configuration pallet. Specifically, that means that the Balance of the Currency implementation should be the same as Balance as used in the Configuration.

type WeightInfo: WeightInfo

Something that provides the weight of this pallet.