Enum emulated_integration_tests_common::impls::v3::prelude::opaque::Junction

pub enum Junction {
    Parachain(u32),
    AccountId32 {
        network: Option<NetworkId>,
        id: [u8; 32],
    },
    AccountIndex64 {
        network: Option<NetworkId>,
        index: u64,
    },
    AccountKey20 {
        network: Option<NetworkId>,
        key: [u8; 20],
    },
    PalletInstance(u8),
    GeneralIndex(u128),
    GeneralKey {
        length: u8,
        data: [u8; 32],
    },
    OnlyChild,
    Plurality {
        id: BodyId,
        part: BodyPart,
    },
    GlobalConsensus(NetworkId),
}

A single item in a path to describe the relative location of a consensus system.

Each item assumes a pre-existing location as its context and is defined in terms of it.

Variants

Parachain(u32)

An indexed parachain belonging to and operated by the context.

Generally used when the context is a Polkadot Relay-chain.

AccountId32

Fields

network: Option<NetworkId>

id: [u8; 32]

A 32-byte identifier for an account of a specific network that is respected as a sovereign endpoint within the context.

Generally used when the context is a Substrate-based chain.

AccountIndex64

Fields

network: Option<NetworkId>

index: u64

An 8-byte index for an account of a specific network that is respected as a sovereign endpoint within the context.

May be used when the context is a Frame-based chain and includes e.g. an indices pallet.

AccountKey20

Fields

network: Option<NetworkId>

key: [u8; 20]

A 20-byte identifier for an account of a specific network that is respected as a sovereign endpoint within the context.

May be used when the context is an Ethereum or Bitcoin chain or smart-contract.

PalletInstance(u8)

An instanced, indexed pallet that forms a constituent part of the context.

Generally used when the context is a Frame-based chain.

GeneralIndex(u128)

A non-descript index within the context location.

Usage will vary widely owing to its generality.

NOTE: Try to avoid using this and instead use a more specific item.

GeneralKey

Fields

length: u8

data: [u8; 32]

A nondescript array datum, 32 bytes, acting as a key within the context location.

Usage will vary widely owing to its generality.

NOTE: Try to avoid using this and instead use a more specific item.

OnlyChild

The unambiguous child.

Not currently used except as a fallback when deriving context.

Plurality

Fields

id: BodyId

part: BodyPart

A pluralistic body existing within consensus.

Typical to be used to represent a governance origin of a chain, but could in principle be used to represent things such as multisigs also.

GlobalConsensus(NetworkId)

A global network capable of externalizing its own consensus. This is not generally meaningful outside of the universal level.

Implementations

impl Junction

pub const fn into_location(self) -> MultiLocation

Convert self into a MultiLocation containing 0 parents.

Similar to Into::into, except that this method can be used in a const evaluation context.

pub const fn into_exterior(self, n: u8) -> MultiLocation

Convert self into a MultiLocation containing n parents.

Similar to Self::into_location, with the added ability to specify the number of parent junctions.

pub const fn into_versioned(self) -> VersionedLocation

Convert self into a VersionedLocation containing 0 parents.

Similar to Into::into, except that this method can be used in a const evaluation context.

pub fn remove_network_id(&mut self)

Remove the NetworkId value.