emulated_integration_tests_common

Module macros

Re-exports

pub use crate::ASSETS_PALLET_ID;

pub use crate::USDT_ID;

pub use paste;

pub use pallet_asset_conversion;

pub use pallet_assets;

pub use pallet_balances;

pub use pallet_message_queue;

pub use pallet_whitelist;

pub use pallet_xcm;

pub use xcm::prelude::AliasOrigin;

pub use xcm::prelude::All;

pub use xcm::prelude::AllCounted;

pub use xcm::prelude::BuyExecution;

pub use xcm::prelude::DepositAsset;

pub use xcm::prelude::ExpectTransactStatus;

pub use xcm::prelude::Fungible;

pub use xcm::prelude::GeneralIndex;

pub use xcm::prelude::Here;

pub use xcm::prelude::InitiateTransfer;

pub use xcm::prelude::PalletInstance;

pub use xcm::prelude::Parachain;

pub use xcm::prelude::PayFees;

pub use xcm::prelude::RefundSurplus;

pub use xcm::prelude::Transact;

pub use xcm::prelude::Unlimited;

pub use xcm::prelude::Wild;

pub use xcm::prelude::WithdrawAsset;

pub use asset_test_utils;

pub use cumulus_pallet_xcmp_queue;

Macros

assert_expected_events

assert_ok

Panic if an expression doesn’t evaluate to Ok.

Structs

AccountId32

An opaque 32-byte cryptographic identifier.

Asset

Either an amount of a single fungible asset, or a single well-identified non-fungible asset.

AssetId

Location to identify an asset.

Assets

A Vec of Assets.

BoundedVec

A bounded vector.

Dmp

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Location

A relative path between state-bearing consensus systems.

Parent

A unit struct which can be converted into a Location of parents value 1.

TestArgs

Default Args provided by xcm-emulator to be stored in a Test instance

TestContext

Auxiliar struct to help creating a new Test instance

Weight

Xcm

XcmContext

Contextual data pertaining to a specific list of XCM instructions.

Enums

AssetTransferFilter

Matches assets based on inner AssetFilter and tags them for a specific type of asset transfer. Please note: the transfer type is specific to each particular (asset, source, dest) combination, so it should always be built in the context of source after knowing dest.

Junction

A single item in a path to describe the relative location of a consensus system.

MaybeErrorCode

OriginKind

Basically just the XCM (more general) version of ParachainDispatchOrigin.

Outcome

Outcome of an XCM execution.

RawOrigin

Origin for the System pallet.

TransferType

Specify which type of asset transfer is required for a particular (asset, dest) combination.

VersionedAssetId

A single version’s AssetId value, together with its version code.

VersionedAssets

A single MultiAssets value, together with its version code.

VersionedLocation

A single Location value, together with its version code.

VersionedXcm

A single XCM message, together with its version code.

WeightLimit

An optional weight limit.

XcmPaymentApiError

Constants

XCM_VERSION

Traits

Chain

Currency

Abstraction over a fungible assets system.

Dispatchable

A lazy call (module function and argument values) that can be executed via its dispatch method.

DropAssets

Define a handler for when some non-empty AssetsInHolding value should be dropped.

DryRunApiV2

API for dry-running extrinsics and XCM programs to get the programs that need to be passed to the fees API.

GetDispatchInfo

A Dispatchable function (aka transaction) that can carry some static information along with it, using the #[pallet::weight] attribute.

Inspect

Trait for providing balance-inspection access to a set of named fungible assets.

Mutate

Trait for providing a basic fungible asset.

Para

RelayChain

TestExt

WeightToFee

A trait that describes the weight to fee calculation.

XcmPaymentApiV1

A trait of XCM payment API.

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.