Module emulated_integration_tests_common::macros

Re-exports

pub use paste;

pub use pallet_assets;

pub use pallet_balances;

pub use pallet_message_queue;

pub use pallet_xcm;

pub use xcm::prelude::AccountId32;

pub use xcm::prelude::All;

pub use xcm::prelude::BuyExecution;

pub use xcm::prelude::DepositAsset;

pub use xcm::prelude::ExpectTransactStatus;

pub use xcm::prelude::Fungible;

pub use xcm::prelude::Here;

pub use xcm::prelude::RefundSurplus;

pub use xcm::prelude::Transact;

pub use xcm::prelude::Unlimited;

pub use xcm::prelude::WithdrawAsset;

pub use asset_test_utils;

pub use cumulus_pallet_xcmp_queue;

Structs

Asset

Either an amount of a single fungible asset, or a single well-identified non-fungible asset.

AssetId

Location to identify an asset.

Location

A relative path between state-bearing consensus systems.

Weight

Xcm

Enums

MaybeErrorCode

OriginKind

Basically just the XCM (more general) version of ParachainDispatchOrigin.

VersionedXcm

A single XCM message, together with its version code.

WeightLimit

An optional weight limit.

Traits

Chain

WeightToFee

A trait that describes the weight to fee calculation.

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

V3Location

Type alias for a better transition to V4.